refactor: 카운트하는 함수들 Count클래스로 이동
